requirements.txt
aenum==3.1.15
affine==2.4.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
dataclasses-json==0.6.7
geopandas==1.0.1
idna==3.10
marshmallow==3.26.1
mypy-extensions==1.0.0
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
rasterio==1.4.3
requests==2.32.3
requests-oauthlib==2.0.0
sentinelhub==3.11.1
shapely==2.0.7
six==1.17.0
tifffile==2025.3.13
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
utm==0.8.1
